feat: adding missed implementation on article reporter relation
