+++ core/src/main/java/com/bbolab/gaonna/core/domain/article/Article.java
@@ -2,7 +2,6 @@ package com.bbolab.gaonna.core.domain.article;
 
 import com.bbolab.gaonna.core.domain.member.MemberArticleComment;
 import com.bbolab.gaonna.core.domain.member.MemberArticleLike;
-import com.bbolab.gaonna.core.domain.member.MemberArticleReport;
 import com.bbolab.gaonna.core.domain.quest.Quest;
 import com.bbolab.gaonna.core.domain.report.ArticleReport;
 import lombok.Builder;
