--- core/src/main/java/com/bbolab/gaonna/core/domain/quest/QuestRoleType.java
@@ -1,5 +1,0 @@
-package com.bbolab.gaonna.core.domain.quest;
-
-public enum QuestRoleType {
-    REQUESTOR, PERFORMER
-}
